backend/requirements.txt
# SmartBot Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
openai==1.10.0
anthropic==0.18.0
httpx==0.26.0
sentence-transformers==2.3.1
langgraph==0.0.20
celery==5.3.6
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
stripe==7.10.0
